gateway/requirements.txt
fastapi==0.95.*
uvicorn==0.21.*
grpcio==1.48.*
protobuf==3.20.*
grpcio-tools==1.48.*
pika==1.3.*
